+++ PrintingConfigTests/NestingTests.cs
@@ -3,12 +3,12 @@ using FluentAssertions;
 using NUnit.Framework;
 using ObjectPrinting;
 using ObjectPrinting.Configuration;
-using Tests.TestingModels;
+using PrintingConfigTests.TestingModels;
 
-namespace Tests
+namespace PrintingConfigTests
 {
     [TestFixture]
-    public class ObjectPrinterNestingTests
+    public class NestingTests
     {
         private string result;
