+++ PrintingConfigTests/TestingModels/TestingPropertiesClass.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Tests.TestingModels
+namespace PrintingConfigTests.TestingModels
 {
     public class TestingPropertiesClass
     {
